pyproject.toml
[project]
name = "vmagic"
version = "0.1.0"
description = "High-fidelity image vectorization library"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "VMagic Contributors"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]
keywords = ["vectorization", "svg", "image-processing", "computer-vision"]

dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "torch>=2.0.0",
    "Pillow>=10.0.0",
    "scikit-image>=0.21.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "vtracer>=0.6.0",
    "cairosvg>=2.7.0",
]

[project.optional-dependencies]
sam = [
    "segment-anything>=1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[project.scripts]
vmagic = "vmagic.cli:main"

[project.urls]
Homepage = "https://github.com/vmagic/vmagic"
Documentation = "https://github.com/vmagic/vmagic#readme"
Repository = "https://github.com/vmagic/vmagic"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["vmagic*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
